init engine process

+++ database/migrations/2020_07_24_042607_create_engines_table.php
@@ -17,7 +17,7 @@ class CreateEnginesTable extends Migration
             $table->id();
             $table->unsignedBigInteger('category_id');
             $table->string('name');
-            $table->integer('capacity');
+            $table->float('capacity');
             $table->timestamps();
         });
     }
